Add files via upload
Introduction to the problem

In Module 1, we are going to get familiar with pandas, the python module which is used to process and analyze data. Processing could include removing unknown values from the data or replacing unknown values with values that make sense, maybe 0.  Analyzing the data could include finding out the trend of a stock price,